chore: downgrade Management API version to /v1alpha
