+++ e2e-tests/api-tests/src/test/java/org/eclipse/edc/identityhub/tests/KeyPairResourceApiEndToEndTest.java
@@ -76,7 +76,7 @@ public class KeyPairResourceApiEndToEndTest extends ManagementApiEndToEndTest {
         RUNTIME_CONFIGURATION.getManagementEndpoint().baseRequest()
                 .contentType(JSON)
                 .header(new Header("x-api-key", user2Token))
-                .get("/v1/participants/%s/keypairs/%s".formatted(toBase64(user1), key))
+                .get("/v1alpha/participants/%s/keypairs/%s".formatted(toBase64(user1), key))
                 .then()
                 .log().ifValidationFails()
                 .statusCode(403)
@@ -95,7 +95,7 @@ public class KeyPairResourceApiEndToEndTest extends ManagementApiEndToEndTest {
                 .allSatisfy(t -> RUNTIME_CONFIGURATION.getManagementEndpoint().baseRequest()
                         .contentType(JSON)
                         .header(new Header("x-api-key", t))
-                        .get("/v1/participants/%s/keypairs/%s".formatted(toBase64(user1), key))
+                        .get("/v1alpha/participants/%s/keypairs/%s".formatted(toBase64(user1), key))
                         .then()
                         .log().ifValidationFails()
                         .statusCode(200)
@@ -122,7 +122,7 @@ public class KeyPairResourceApiEndToEndTest extends ManagementApiEndToEndTest {
         var res = RUNTIME_CONFIGURATION.getManagementEndpoint().baseRequest()
                 .contentType(JSON)
                 .header(new Header("x-api-key", user2Token))
-                .get("/v1/participants/%s/keypairs".formatted(toBase64(user1)))
+                .get("/v1alpha/participants/%s/keypairs".formatted(toBase64(user1)))
                 .then()
                 .log().ifValidationFails()
                 .statusCode(200)
@@ -143,7 +143,7 @@ public class KeyPairResourceApiEndToEndTest extends ManagementApiEndToEndTest {
                 .allSatisfy(t -> RUNTIME_CONFIGURATION.getManagementEndpoint().baseRequest()
                         .contentType(JSON)
                         .header(new Header("x-api-key", t))
-                        .get("/v1/participants/%s/keypairs".formatted(toBase64(user1)))
+                        .get("/v1alpha/participants/%s/keypairs".formatted(toBase64(user1)))
                         .then()
                         .log().ifValidationFails()
                         .statusCode(200)
@@ -167,7 +167,7 @@ public class KeyPairResourceApiEndToEndTest extends ManagementApiEndToEndTest {
                             .contentType(JSON)
                             .header(new Header("x-api-key", t))
                             .body(keyDesc)
-                            .put("/v1/participants/%s/keypairs".formatted(toBase64(user1)))
+                            .put("/v1alpha/participants/%s/keypairs".formatted(toBase64(user1)))
                             .then()
                             .log().ifValidationFails()
                             .statusCode(204)
@@ -198,7 +198,7 @@ public class KeyPairResourceApiEndToEndTest extends ManagementApiEndToEndTest {
                 .contentType(JSON)
                 .header(new Header("x-api-key", token2))
                 .body(keyDesc)
-                .put("/v1/participants/%s/keypairs".formatted(toBase64(user1)))
+                .put("/v1alpha/participants/%s/keypairs".formatted(toBase64(user1)))
                 .then()
                 .log().ifValidationFails()
                 .statusCode(403)
@@ -233,7 +233,7 @@ public class KeyPairResourceApiEndToEndTest extends ManagementApiEndToEndTest {
                             .contentType(JSON)
                             .header(new Header("x-api-key", t))
                             .body(keyDesc)
-                            .post("/v1/participants/%s/keypairs/%s/rotate".formatted(toBase64(user1), keyId))
+                            .post("/v1alpha/participants/%s/keypairs/%s/rotate".formatted(toBase64(user1), keyId))
                             .then()
                             .log().ifValidationFails()
                             .statusCode(204)
@@ -275,7 +275,7 @@ public class KeyPairResourceApiEndToEndTest extends ManagementApiEndToEndTest {
                 .contentType(JSON)
                 .header(new Header("x-api-key", token2))
                 .body(keyDesc)
-                .post("/v1/participants/%s/keypairs/%s/rotate".formatted(user1, keyId))
+                .post("/v1alpha/participants/%s/keypairs/%s/rotate".formatted(user1, keyId))
                 .then()
                 .log().ifValidationFails()
                 .statusCode(403)
@@ -305,7 +305,7 @@ public class KeyPairResourceApiEndToEndTest extends ManagementApiEndToEndTest {
                             .contentType(JSON)
                             .header(new Header("x-api-key", t))
                             .body(keyDesc)
-                            .post("/v1/participants/%s/keypairs/%s/revoke".formatted(toBase64(user1), keyId))
+                            .post("/v1alpha/participants/%s/keypairs/%s/revoke".formatted(toBase64(user1), keyId))
                             .then()
                             .log().ifValidationFails()
                             .statusCode(204)
@@ -332,7 +332,7 @@ public class KeyPairResourceApiEndToEndTest extends ManagementApiEndToEndTest {
                 .contentType(JSON)
                 .header(new Header("x-api-key", token2))
                 .body(keyDesc)
-                .post("/v1/participants/%s/keypairs/%s/revoke".formatted(toBase64(user1), keyId))
+                .post("/v1alpha/participants/%s/keypairs/%s/revoke".formatted(toBase64(user1), keyId))
                 .then()
                 .log().ifValidationFails()
                 .statusCode(403)
@@ -350,7 +350,7 @@ public class KeyPairResourceApiEndToEndTest extends ManagementApiEndToEndTest {
         var found = RUNTIME_CONFIGURATION.getManagementEndpoint().baseRequest()
                 .contentType(JSON)
                 .header(new Header("x-api-key", superUserKey))
-                .get("/v1/keypairs")
+                .get("/v1alpha/keypairs")
                 .then()
                 .log().ifValidationFails()
                 .statusCode(200)
@@ -369,7 +369,7 @@ public class KeyPairResourceApiEndToEndTest extends ManagementApiEndToEndTest {
         var found = RUNTIME_CONFIGURATION.getManagementEndpoint().baseRequest()
                 .contentType(JSON)
                 .header(new Header("x-api-key", superUserKey))
-                .get("/v1/keypairs?offset=2&limit=4")
+                .get("/v1alpha/keypairs?offset=2&limit=4")
                 .then()
                 .log().ifValidationFails()
                 .statusCode(200)
@@ -388,7 +388,7 @@ public class KeyPairResourceApiEndToEndTest extends ManagementApiEndToEndTest {
         var found = RUNTIME_CONFIGURATION.getManagementEndpoint().baseRequest()
                 .contentType(JSON)
                 .header(new Header("x-api-key", superUserKey))
-                .get("/v1/keypairs")
+                .get("/v1alpha/keypairs")
                 .then()
                 .log().ifValidationFails()
                 .statusCode(200)
@@ -408,7 +408,7 @@ public class KeyPairResourceApiEndToEndTest extends ManagementApiEndToEndTest {
         RUNTIME_CONFIGURATION.getManagementEndpoint().baseRequest()
                 .contentType(JSON)
                 .header(new Header("x-api-key", attackerToken))
-                .get("/v1/keypairs")
+                .get("/v1alpha/keypairs")
                 .then()
                 .log().ifValidationFails()
                 .statusCode(403);
@@ -426,7 +426,7 @@ public class KeyPairResourceApiEndToEndTest extends ManagementApiEndToEndTest {
                     RUNTIME_CONFIGURATION.getManagementEndpoint().baseRequest()
                             .contentType(JSON)
                             .header(new Header("x-api-key", t))
-                            .post("/v1/participants/%s/keypairs/%s/activate".formatted(toBase64(user1), keyId))
+                            .post("/v1alpha/participants/%s/keypairs/%s/activate".formatted(toBase64(user1), keyId))
                             .then()
                             .log().ifValidationFails()
                             .statusCode(204)
@@ -448,7 +448,7 @@ public class KeyPairResourceApiEndToEndTest extends ManagementApiEndToEndTest {
         RUNTIME_CONFIGURATION.getManagementEndpoint().baseRequest()
                 .contentType(JSON)
                 .header(new Header("x-api-key", attackerToken))
-                .post("/v1/participants/%s/keypairs/%s/activate".formatted(toBase64(user1), keyId))
+                .post("/v1alpha/participants/%s/keypairs/%s/activate".formatted(toBase64(user1), keyId))
                 .then()
                 .log().ifValidationFails()
                 .statusCode(403)
@@ -469,7 +469,7 @@ public class KeyPairResourceApiEndToEndTest extends ManagementApiEndToEndTest {
         RUNTIME_CONFIGURATION.getManagementEndpoint().baseRequest()
                 .contentType(JSON)
                 .header(new Header("x-api-key", token))
-                .post("/v1/participants/%s/keypairs/%s/revoke".formatted(toBase64(user1), keyId))
+                .post("/v1alpha/participants/%s/keypairs/%s/revoke".formatted(toBase64(user1), keyId))
                 .then()
                 .log().ifValidationFails()
                 .statusCode(204)
@@ -479,7 +479,7 @@ public class KeyPairResourceApiEndToEndTest extends ManagementApiEndToEndTest {
         RUNTIME_CONFIGURATION.getManagementEndpoint().baseRequest()
                 .contentType(JSON)
                 .header(new Header("x-api-key", token))
-                .post("/v1/participants/%s/keypairs/%s/activate".formatted(toBase64(user1), keyId))
+                .post("/v1alpha/participants/%s/keypairs/%s/activate".formatted(toBase64(user1), keyId))
                 .then()
                 .log().ifValidationFails()
                 .statusCode(400)
