+++ e2e-tests/api-tests/src/test/java/org/eclipse/edc/identityhub/tests/ParticipantContextApiEndToEndTest.java
@@ -56,7 +56,7 @@ public class ParticipantContextApiEndToEndTest extends ManagementApiEndToEndTest
 
         var su = RUNTIME_CONFIGURATION.getManagementEndpoint().baseRequest()
                 .header(new Header("x-api-key", apikey))
-                .get("/v1/participants/" + toBase64(SUPER_USER))
+                .get("/v1alpha/participants/" + toBase64(SUPER_USER))
                 .then()
                 .statusCode(200)
                 .extract().body().as(ParticipantContext.class);
@@ -85,7 +85,7 @@ public class ParticipantContextApiEndToEndTest extends ManagementApiEndToEndTest
         RUNTIME_CONFIGURATION.getManagementEndpoint().baseRequest()
                 .header(new Header("x-api-key", apiToken1))
                 .contentType(ContentType.JSON)
-                .get("/v1/participants/" + toBase64(user2))
+                .get("/v1alpha/participants/" + toBase64(user2))
                 .then()
                 .log().ifValidationFails()
                 .statusCode(403);
@@ -103,7 +103,7 @@ public class ParticipantContextApiEndToEndTest extends ManagementApiEndToEndTest
                 .header(new Header("x-api-key", apikey))
                 .contentType(ContentType.JSON)
                 .body(manifest)
-                .post("/v1/participants/")
+                .post("/v1alpha/participants/")
                 .then()
                 .log().ifError()
                 .statusCode(anyOf(equalTo(200), equalTo(204)))
@@ -134,7 +134,7 @@ public class ParticipantContextApiEndToEndTest extends ManagementApiEndToEndTest
                 .header(new Header("x-api-key", apikey))
                 .contentType(ContentType.JSON)
                 .body(manifest)
-                .post("/v1/participants/")
+                .post("/v1alpha/participants/")
                 .then()
                 .log().ifError()
                 .statusCode(anyOf(equalTo(200), equalTo(204)))
@@ -168,7 +168,7 @@ public class ParticipantContextApiEndToEndTest extends ManagementApiEndToEndTest
                 .header(new Header("x-api-key", apiToken))
                 .contentType(ContentType.JSON)
                 .body(manifest)
-                .post("/v1/participants/")
+                .post("/v1alpha/participants/")
                 .then()
                 .log().ifError()
                 .statusCode(403)
@@ -190,7 +190,7 @@ public class ParticipantContextApiEndToEndTest extends ManagementApiEndToEndTest
                 .header(new Header("x-api-key", createTokenFor(principal)))
                 .contentType(ContentType.JSON)
                 .body(manifest)
-                .post("/v1/participants/")
+                .post("/v1alpha/participants/")
                 .then()
                 .log().ifError()
                 .statusCode(401)
@@ -217,7 +217,7 @@ public class ParticipantContextApiEndToEndTest extends ManagementApiEndToEndTest
         RUNTIME_CONFIGURATION.getManagementEndpoint().baseRequest()
                 .header(new Header("x-api-key", superUserKey))
                 .contentType(ContentType.JSON)
-                .post("/v1/participants/%s/state?isActive=true".formatted(toBase64(participantId)))
+                .post("/v1alpha/participants/%s/state?isActive=true".formatted(toBase64(participantId)))
                 .then()
                 .log().ifError()
                 .statusCode(204);
@@ -243,7 +243,7 @@ public class ParticipantContextApiEndToEndTest extends ManagementApiEndToEndTest
         RUNTIME_CONFIGURATION.getManagementEndpoint().baseRequest()
                 .header(new Header("x-api-key", superUserKey))
                 .contentType(ContentType.JSON)
-                .delete("/v1/participants/%s".formatted(toBase64(participantId)))
+                .delete("/v1alpha/participants/%s".formatted(toBase64(participantId)))
                 .then()
                 .log().ifError()
                 .statusCode(204);
@@ -261,7 +261,7 @@ public class ParticipantContextApiEndToEndTest extends ManagementApiEndToEndTest
                 .allSatisfy(t -> RUNTIME_CONFIGURATION.getManagementEndpoint().baseRequest()
                         .header(new Header("x-api-key", t))
                         .contentType(ContentType.JSON)
-                        .post("/v1/participants/%s/token".formatted(toBase64(participantId)))
+                        .post("/v1alpha/participants/%s/token".formatted(toBase64(participantId)))
                         .then()
                         .log().ifError()
                         .statusCode(200)
@@ -278,7 +278,7 @@ public class ParticipantContextApiEndToEndTest extends ManagementApiEndToEndTest
                 .header(new Header("x-api-key", superUserKey))
                 .contentType(ContentType.JSON)
                 .body(List.of("role1", "role2", "admin"))
-                .put("/v1/participants/%s/roles".formatted(toBase64(participantId)))
+                .put("/v1alpha/participants/%s/roles".formatted(toBase64(participantId)))
                 .then()
                 .log().ifError()
                 .statusCode(204);
@@ -296,7 +296,7 @@ public class ParticipantContextApiEndToEndTest extends ManagementApiEndToEndTest
                 .header(new Header("x-api-key", userToken))
                 .contentType(ContentType.JSON)
                 .body(List.of(role))
-                .put("/v1/participants/%s/roles".formatted(toBase64(participantId)))
+                .put("/v1alpha/participants/%s/roles".formatted(toBase64(participantId)))
                 .then()
                 .log().ifError()
                 .statusCode(403);
@@ -313,7 +313,7 @@ public class ParticipantContextApiEndToEndTest extends ManagementApiEndToEndTest
         var found = RUNTIME_CONFIGURATION.getManagementEndpoint().baseRequest()
                 .contentType(JSON)
                 .header(new Header("x-api-key", superUserKey))
-                .get("/v1/participants")
+                .get("/v1alpha/participants")
                 .then()
                 .log().ifValidationFails()
                 .statusCode(200)
@@ -332,7 +332,7 @@ public class ParticipantContextApiEndToEndTest extends ManagementApiEndToEndTest
         var found = RUNTIME_CONFIGURATION.getManagementEndpoint().baseRequest()
                 .contentType(JSON)
                 .header(new Header("x-api-key", superUserKey))
-                .get("/v1/participants?offset=2&limit=4")
+                .get("/v1alpha/participants?offset=2&limit=4")
                 .then()
                 .log().ifValidationFails()
                 .statusCode(200)
@@ -351,7 +351,7 @@ public class ParticipantContextApiEndToEndTest extends ManagementApiEndToEndTest
         var found = RUNTIME_CONFIGURATION.getManagementEndpoint().baseRequest()
                 .contentType(JSON)
                 .header(new Header("x-api-key", superUserKey))
-                .get("/v1/participants")
+                .get("/v1alpha/participants")
                 .then()
                 .log().ifValidationFails()
                 .statusCode(200)
@@ -371,7 +371,7 @@ public class ParticipantContextApiEndToEndTest extends ManagementApiEndToEndTest
         RUNTIME_CONFIGURATION.getManagementEndpoint().baseRequest()
                 .contentType(JSON)
                 .header(new Header("x-api-key", attackerToken))
-                .get("/v1/participants")
+                .get("/v1alpha/participants")
                 .then()
                 .log().ifValidationFails()
                 .statusCode(403);
