+++ extensions/api/management-api/did-api/src/main/java/org/eclipse/edc/identityhub/api/didmanagement/DidManagementApiExtension.java
@@ -16,8 +16,8 @@ package org.eclipse.edc.identityhub.api.didmanagement;
 
 import org.eclipse.edc.identithub.spi.did.DidDocumentService;
 import org.eclipse.edc.identithub.spi.did.model.DidResource;
-import org.eclipse.edc.identityhub.api.didmanagement.v1.DidManagementApiController;
-import org.eclipse.edc.identityhub.api.didmanagement.v1.GetAllDidsApiController;
+import org.eclipse.edc.identityhub.api.didmanagement.v1.unstable.DidManagementApiController;
+import org.eclipse.edc.identityhub.api.didmanagement.v1.unstable.GetAllDidsApiController;
 import org.eclipse.edc.identityhub.spi.AuthorizationService;
 import org.eclipse.edc.identityhub.spi.ManagementApiConfiguration;
 import org.eclipse.edc.runtime.metamodel.annotation.Extension;
