+++ extensions/api/management-api/keypair-api/src/main/java/org/eclipse/edc/identityhub/api/keypair/v1/unstable/GetAllKeyPairsApi.java
@@ -12,7 +12,7 @@
  *
  */
 
-package org.eclipse.edc.identityhub.api.didmanagement.v1;
+package org.eclipse.edc.identityhub.api.keypair.v1.unstable;
 
 import io.swagger.v3.oas.annotations.OpenAPIDefinition;
 import io.swagger.v3.oas.annotations.Operation;
@@ -23,30 +23,30 @@ import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.tags.Tag;
-import org.eclipse.edc.iam.did.spi.document.DidDocument;
+import org.eclipse.edc.identityhub.spi.keypair.model.KeyPairResource;
 import org.eclipse.edc.web.spi.ApiErrorDetail;
 
 import java.util.Collection;
 
-@OpenAPIDefinition(
-        info = @Info(description = "This is the Management API for DID documents", title = "DID Management API", version = "1"))
-@Tag(name = "DID")
-public interface GetAllDidsApi {
+
+@OpenAPIDefinition(info = @Info(description = "This is the Management API for KeyPairResources", title = "KeyPairResources Management API", version = "1"))
+@Tag(name = "Key Pairs")
+public interface GetAllKeyPairsApi {
 
 
-    @Operation(description = "Get all DID documents across all Participant Contexts. Requires elevated access.",
-            operationId = "getAllDids",
+    @Operation(description = "Get all KeyPair resources across all Participant Contexts. Requires elevated access.",
+            operationId = "getAllKeyPairs",
             parameters = {
                     @Parameter(name = "offset", description = "the paging offset. defaults to 0"),
-                    @Parameter(name = "limit", description = "the page size. defaults to 50")},
+                    @Parameter(name = "limit", description = "the page size. defaults to 50") },
             responses = {
-                    @ApiResponse(responseCode = "200", description = "The list of DID Documents.",
-                            content = @Content(array = @ArraySchema(schema = @Schema(implementation = DidDocument.class)))),
+                    @ApiResponse(responseCode = "200", description = "The list of KeyPair resources.",
+                            content = @Content(array = @ArraySchema(schema = @Schema(implementation = KeyPairResource.class)))),
                     @ApiResponse(responseCode = "401", description = "The request could not be completed, because either the authentication was missing or was not valid.",
                             content = @Content(array = @ArraySchema(schema = @Schema(implementation = ApiErrorDetail.class)), mediaType = "application/json")),
                     @ApiResponse(responseCode = "400", description = "The query was malformed or was not understood by the server.",
                             content = @Content(array = @ArraySchema(schema = @Schema(implementation = ApiErrorDetail.class)), mediaType = "application/json")),
             }
     )
-    Collection<DidDocument> getAllDids(Integer offset, Integer limit);
+    Collection<KeyPairResource> getAllKeyPairs(Integer offset, Integer limit);
 }
